third_party/WebKit/LayoutTests/vr/latest/vrDevice_requestSession_exclusive_unsupported.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../resources/mock-vr-service.js"></script>
<script src="../resources/test-constants.js"></script>
<canvas id="webgl-canvas"></canvas>
<script src="../resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();
vr_test( (t) => {
  return navigator.vr.requestDevice().then( (magicWindowOnly) => {
    runWithUserGesture( () => {

      magicWindowOnly.requestSession({ exclusive: true }).then(() => {
        assert_unreached("requestSession resolved unexpectedly");
      }).catch((e) => {
        assert_equals(e.name, "NotSupportedError");
        t.done();
      })

      //TODO: Add magicWindowOnly.requestSession({ exclusive: false })

    });
  }, (err) => {
    t.step( () => {
      assert_unreached("requestDevice rejected");
    });
  });
}, [fakeDisplays["FakeMagicWindowOnly"]],
"supportsSession properly creates or rejects exclusive sessions");

</script>
